refactor: split shell presentation selectors
