+++ apps/desktop/src/components/core/ComposerPanel.tsx
@@ -7,7 +7,7 @@ import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
-import { shortPubkey } from '@/shell/selectors';
+import { shortPubkey } from '@/shell/presentation';
 
 import { AuthorAvatar } from './AuthorAvatar';
 import { ComposerDraftPreviewList } from './ComposerDraftPreviewList';
