+++ apps/desktop/src/shell/storeSelectors.ts
@@ -1,0 +1,164 @@
+import type { DesktopShellStore } from '@/shell/store';
+
+/// useDesktopShellActions.ts が読むフィールド(購読スライス。WP-H6 PR2)。
+export const selectShellActionsSlice = (s: DesktopShellStore) => ({
+  activeTopic: s.activeTopic,
+  bookmarkedPosts: s.bookmarkedPosts,
+  channelAudienceInput: s.channelAudienceInput,
+  channelLabelInput: s.channelLabelInput,
+  communityNodeInput: s.communityNodeInput,
+  composeChannelByTopic: s.composeChannelByTopic,
+  composer: s.composer,
+  discoverySeedInput: s.discoverySeedInput,
+  draftMediaItems: s.draftMediaItems,
+  gameDescription: s.gameDescription,
+  gameDrafts: s.gameDrafts,
+  gameParticipantsInput: s.gameParticipantsInput,
+  gameRoomsByTopic: s.gameRoomsByTopic,
+  gameTitle: s.gameTitle,
+  inviteTokenInput: s.inviteTokenInput,
+  joinedChannelsByTopic: s.joinedChannelsByTopic,
+  liveDescription: s.liveDescription,
+  liveTitle: s.liveTitle,
+  localProfile: s.localProfile,
+  peerTicket: s.peerTicket,
+  profileDraft: s.profileDraft,
+  replyTarget: s.replyTarget,
+  repostTarget: s.repostTarget,
+  selectedAuthorPubkey: s.selectedAuthorPubkey,
+  selectedChannelIdByTopic: s.selectedChannelIdByTopic,
+  selectedThread: s.selectedThread,
+  shellChromeState: s.shellChromeState,
+  syncStatus: s.syncStatus,
+  topicInput: s.topicInput,
+  trackedTopics: s.trackedTopics,
+
+});
+
+/// useDesktopShellData.ts が読むフィールド(購読スライス。WP-H6 PR2)。
+export const selectShellDataSlice = (s: DesktopShellStore) => ({
+  activeTopic: s.activeTopic,
+  bookmarkedReactionAssets: s.bookmarkedReactionAssets,
+  directMessageTimelineByPeer: s.directMessageTimelineByPeer,
+  gameRoomsByTopic: s.gameRoomsByTopic,
+  joinedChannelsByTopic: s.joinedChannelsByTopic,
+  knownAuthorsByPubkey: s.knownAuthorsByPubkey,
+  localProfile: s.localProfile,
+  mediaObjectUrls: s.mediaObjectUrls,
+  notifications: s.notifications,
+  ownedReactionAssets: s.ownedReactionAssets,
+  profileTimeline: s.profileTimeline,
+  publicTimelinesByTopic: s.publicTimelinesByTopic,
+  recentReactions: s.recentReactions,
+  selectedAuthorPubkey: s.selectedAuthorPubkey,
+  selectedAuthorTimeline: s.selectedAuthorTimeline,
+  selectedChannelIdByTopic: s.selectedChannelIdByTopic,
+  selectedDirectMessagePeerPubkey: s.selectedDirectMessagePeerPubkey,
+  selectedThread: s.selectedThread,
+  shellChromeState: s.shellChromeState,
+  syncStatus: s.syncStatus,
+  thread: s.thread,
+  timelineScopeByTopic: s.timelineScopeByTopic,
+  timelinesByKey: s.timelinesByKey,
+  trackedTopics: s.trackedTopics,
+
+});
+
+/// useDesktopShellRouting.ts が読むフィールド(購読スライス。WP-H6 PR2)。
+export const selectShellRoutingSlice = (s: DesktopShellStore) => ({
+  activeTopic: s.activeTopic,
+  channelPanelStateByTopic: s.channelPanelStateByTopic,
+  directMessagePaneOpen: s.directMessagePaneOpen,
+  focusedObjectId: s.focusedObjectId,
+  gamePanelStateByTopic: s.gamePanelStateByTopic,
+  gameRoomsByTopic: s.gameRoomsByTopic,
+  joinedChannelsByTopic: s.joinedChannelsByTopic,
+  lastNonNotificationsRoute: s.lastNonNotificationsRoute,
+  livePanelStateByTopic: s.livePanelStateByTopic,
+  liveSessionsByTopic: s.liveSessionsByTopic,
+  selectedAuthor: s.selectedAuthor,
+  selectedAuthorPubkey: s.selectedAuthorPubkey,
+  selectedChannelIdByTopic: s.selectedChannelIdByTopic,
+  selectedDirectMessagePeerPubkey: s.selectedDirectMessagePeerPubkey,
+  selectedGameRoomId: s.selectedGameRoomId,
+  selectedLiveSessionId: s.selectedLiveSessionId,
+  selectedThread: s.selectedThread,
+  shellChromeState: s.shellChromeState,
+  thread: s.thread,
+  trackedTopics: s.trackedTopics,
+    
+});
+
+/// useDesktopShellViewModels.ts が読むフィールド(購読スライス。WP-H6 PR2)。
+export const selectShellViewModelsSlice = (s: DesktopShellStore) => ({
+  activeTopic: s.activeTopic,
+  authorError: s.authorError,
+  bookmarkedPosts: s.bookmarkedPosts,
+  bookmarkedReactionAssets: s.bookmarkedReactionAssets,
+  channelPanelStateByTopic: s.channelPanelStateByTopic,
+  communityNodeConfig: s.communityNodeConfig,
+  communityNodeEditorDirty: s.communityNodeEditorDirty,
+  communityNodeError: s.communityNodeError,
+  communityNodeInput: s.communityNodeInput,
+  communityNodeManifests: s.communityNodeManifests,
+  communityNodeStatuses: s.communityNodeStatuses,
+  composeChannelByTopic: s.composeChannelByTopic,
+  directMessageDraftMediaItems: s.directMessageDraftMediaItems,
+  directMessageStatusByPeer: s.directMessageStatusByPeer,
+  directMessageTimelineByPeer: s.directMessageTimelineByPeer,
+  directMessages: s.directMessages,
+  discoveryConfig: s.discoveryConfig,
+  discoveryEditorDirty: s.discoveryEditorDirty,
+  discoveryError: s.discoveryError,
+  discoverySeedInput: s.discoverySeedInput,
+  draftMediaItems: s.draftMediaItems,
+  error: s.error,
+  gameDrafts: s.gameDrafts,
+  gamePanelStateByTopic: s.gamePanelStateByTopic,
+  gameRoomsByTopic: s.gameRoomsByTopic,
+  joinedChannelsByTopic: s.joinedChannelsByTopic,
+  knownAuthorsByPubkey: s.knownAuthorsByPubkey,
+  livePanelStateByTopic: s.livePanelStateByTopic,
+  livePendingBySessionId: s.livePendingBySessionId,
+  liveSessionsByTopic: s.liveSessionsByTopic,
+  localPeerTicket: s.localPeerTicket,
+  localProfile: s.localProfile,
+  mediaObjectUrls: s.mediaObjectUrls,
+  ownedReactionAssets: s.ownedReactionAssets,
+  peerTicket: s.peerTicket,
+  profileDraft: s.profileDraft,
+  profileTimeline: s.profileTimeline,
+  reactionPanelState: s.reactionPanelState,
+  replyTarget: s.replyTarget,
+  repostTarget: s.repostTarget,
+  selectedAuthor: s.selectedAuthor,
+  selectedAuthorTimeline: s.selectedAuthorTimeline,
+  selectedChannelIdByTopic: s.selectedChannelIdByTopic,
+  selectedDirectMessagePeerPubkey: s.selectedDirectMessagePeerPubkey,
+  selectedThread: s.selectedThread,
+  shellChromeState: s.shellChromeState,
+  socialConnections: s.socialConnections,
+  syncStatus: s.syncStatus,
+  thread: s.thread,
+  timelineScopeByTopic: s.timelineScopeByTopic,
+  timelinesByKey: s.timelinesByKey,
+  trackedTopics: s.trackedTopics,
+  unsupportedVideoManifests: s.unsupportedVideoManifests,
+    
+});
+
+/// DesktopShellPage が読むフィールド(購読スライス。WP-H6 PR2)。
+export const selectShellPageSlice = (s: DesktopShellStore) => ({
+  trackedTopics: s.trackedTopics,
+  activeTopic: s.activeTopic,
+  topicInput: s.topicInput,
+  selectedThread: s.selectedThread,
+  focusedObjectId: s.focusedObjectId,
+  syncStatus: s.syncStatus,
+  selectedAuthorPubkey: s.selectedAuthorPubkey,
+  notifications: s.notifications,
+  notificationStatus: s.notificationStatus,
+  selectedLiveSessionId: s.selectedLiveSessionId,
+  selectedGameRoomId: s.selectedGameRoomId,
+  shellChromeState: s.shellChromeState,
+});
